chore: handleDelete item 내부로 이동
